+++ src/icons/fileIcons.ts
@@ -277,6 +277,9 @@ export const fileIcons: FileIcons = {
         'vsixmanifest',
         'vsix',
         'code-workplace',
+        'code-workspace',
+        'code-profile',
+        'code-snippets',
       ],
     },
     {
@@ -347,12 +350,14 @@ export const fileIcons: FileIcons = {
     { name: 'h', fileExtensions: ['h'] },
     {
       name: 'cpp',
-      fileExtensions: ['cc', 'cpp', 'cxx', 'c++', 'cp', 'mm', 'mii', 'ii'],
+      fileExtensions: ['cc', 'cpp', 'cxx', 'c++', 'cp', 'mii', 'ii'],
     },
     {
       name: 'hpp',
       fileExtensions: ['hh', 'hpp', 'hxx', 'h++', 'hp', 'tcc', 'inl'],
     },
+    { name: 'objective-c', fileExtensions: ['m'] },
+    { name: 'objective-cpp', fileExtensions: ['mm'] },
     { name: 'go', fileExtensions: ['go'] },
     {
       name: 'go-mod',
@@ -601,6 +606,7 @@ export const fileIcons: FileIcons = {
         '.graphqlrc.yml',
         'graphql.config.json',
         'graphql.config.js',
+        'graphql.config.cjs',
         'graphql.config.ts',
         'graphql.config.toml',
         'graphql.config.yaml',
@@ -619,6 +625,7 @@ export const fileIcons: FileIcons = {
       fileNames: [
         '.gitignore',
         '.git-blame-ignore',
+        '.git-blame-ignore-revs',
         '.gitignore-global',
         '.gitignore_global',
         '.gitconfig',
@@ -712,6 +719,8 @@ export const fileIcons: FileIcons = {
         'test.cts',
         'test.mts',
         'ts.snap',
+        'spec-d.ts',
+        'test-d.ts',
       ],
     },
     {
@@ -725,6 +734,8 @@ export const fileIcons: FileIcons = {
         'jsx.snap',
         'cy.jsx',
         'cy.tsx',
+        'spec-d.tsx',
+        'test-d.tsx',
       ],
     },
     {
@@ -880,7 +891,11 @@ export const fileIcons: FileIcons = {
         '.posthtmlrc.yml',
       ],
     },
-    { name: 'todo', fileExtensions: ['todo'] },
+    {
+      name: 'todo',
+      fileExtensions: ['todo'],
+      fileNames: ['todo.md', 'todos.md'],
+    },
     { name: 'coldfusion', fileExtensions: ['cfml', 'cfc', 'lucee', 'cfm'] },
     {
       name: 'cabal',
@@ -1150,6 +1165,7 @@ export const fileIcons: FileIcons = {
         '.env.production.local',
         '.env.staging.local',
         '.env.test.local',
+        '.env.uat',
       ],
     },
     {
@@ -1301,7 +1317,11 @@ export const fileIcons: FileIcons = {
     },
     {
       name: 'conduct',
-      fileNames: ['code_of_conduct.md', 'code_of_conduct.txt'],
+      fileNames: [
+        'code_of_conduct.md',
+        'code_of_conduct.txt',
+        'code_of_conduct',
+      ],
     },
     { name: 'watchman', fileNames: ['.watchmanconfig'] },
     { name: 'aurelia', fileNames: ['aurelia.json'] },
@@ -1473,9 +1493,11 @@ export const fileIcons: FileIcons = {
     {
       name: 'gatsby',
       fileNames: [
-        'gatsby-config.ts',
         'gatsby-config.js',
+        'gatsby-config.mjs',
+        'gatsby-config.ts',
         'gatsby-node.js',
+        'gatsby-node.mjs',
         'gatsby-node.ts',
         'gatsby-browser.js',
         'gatsby-browser.tsx',
@@ -1603,7 +1625,18 @@ export const fileIcons: FileIcons = {
     { name: 'mint', fileExtensions: ['mint'] },
     { name: 'velocity', fileExtensions: ['vm', 'fhtml', 'vtl'] },
     { name: 'godot', fileExtensions: ['gd'] },
-    { name: 'godot-assets', fileExtensions: ['godot', 'tres', 'tscn'] },
+    {
+      name: 'godot-assets',
+      fileExtensions: [
+        'godot',
+        'tres',
+        'tscn',
+        'gdns',
+        'gdnlib',
+        'gdshader',
+        'gdextension',
+      ],
+    },
     {
       name: 'azure-pipelines',
       fileNames: ['azure-pipelines.yml', 'azure-pipelines.yaml'],
@@ -1924,6 +1957,17 @@ export const fileIcons: FileIcons = {
         'vite.config.mts',
       ],
     },
+    {
+      name: 'vitest',
+      fileNames: [
+        'vitest.config.ts',
+        'vitest.config.mts',
+        'vitest.config.cts',
+        'vitest.config.js',
+        'vitest.config.mjs',
+        'vitest.config.cjs',
+      ],
+    },
     { name: 'opa', fileExtensions: ['rego'] },
     { name: 'lerna', fileNames: ['lerna.json'] },
     {
@@ -1993,6 +2037,7 @@ export const fileIcons: FileIcons = {
         'tsconfig.cjs.json',
         'tsconfig.esm.json',
         'tsconfig.mjs.json',
+        'tsconfig.doc.json',
       ],
       fileExtensions: ['tsconfig.json'],
     },
@@ -2120,7 +2165,10 @@ export const fileIcons: FileIcons = {
     },
     { name: 'siyuan', fileExtensions: ['sy'] },
     { name: 'ndst', fileExtensions: ['ndst.yml', 'ndst.yaml', 'ndst.json'] },
-    { name: 'plop', fileNames: ['plopfile.js', 'plopfile.ts'] },
+    {
+      name: 'plop',
+      fileNames: ['plopfile.js', 'plopfile.cjs', 'plopfile.mjs', 'plopfile.ts'],
+    },
     { name: 'tobi', fileExtensions: ['tobi'] },
     { name: 'tobimake', fileNames: ['.tobimake'] },
     { name: 'gleam', fileNames: ['gleam.toml'], fileExtensions: ['gleam'] },
@@ -2149,6 +2197,7 @@ export const fileIcons: FileIcons = {
       fileExtensions: ['cdc'],
     },
     { name: 'caddy', fileNames: ['Caddyfile'] },
+    { name: 'bun', fileNames: ['bun.lockb'], light: true },
     { name: 'antlr', fileExtensions: ['g4'] },
     { name: 'stylable', fileExtensions: ['st.css'] },
     { name: 'pinejs', fileExtensions: ['pine'] },
@@ -2188,5 +2237,10 @@ export const fileIcons: FileIcons = {
         '.cracorc',
       ],
     },
+    {
+      name: 'gamemaker',
+      fileExtensions: ['gml', 'yy', 'yyp', 'yyz'],
+    },
+    { name: 'tldraw', fileExtensions: ['tldr'], light: true },
   ],
 };