+++ src/icons/folderIcons.ts
@@ -25,6 +25,10 @@ export const folderIcons: FolderTheme[] = [
       {
         name: 'folder-images',
         folderNames: [
+          '_images',
+          '_image',
+          '_imgs',
+          '_img',
           'images',
           'image',
           'imgs',
@@ -83,6 +87,8 @@ export const folderIcons: FolderTheme[] = [
       {
         name: 'folder-docs',
         folderNames: [
+          '_post',
+          '_posts',
           'doc',
           'docs',
           'document',
@@ -210,6 +216,7 @@ export const folderIcons: FolderTheme[] = [
       {
         name: 'folder-public',
         folderNames: [
+          '_site',
           'public',
           'www',
           'wwwroot',
@@ -416,6 +423,7 @@ export const folderIcons: FolderTheme[] = [
         name: 'folder-gulp',
         folderNames: [
           'gulp',
+          'gulp-tasks',
           'gulpfile.js',
           'gulpfile.mjs',
           'gulpfile.ts',
@@ -467,7 +475,10 @@ export const folderIcons: FolderTheme[] = [
       { name: 'folder-ios', folderNames: ['ios'] },
       { name: 'folder-upload', folderNames: ['uploads', 'upload'] },
       { name: 'folder-download', folderNames: ['downloads', 'download'] },
-      { name: 'folder-tools', folderNames: ['tools'] },
+      {
+        name: 'folder-tools',
+        folderNames: ['tools', 'toolkit', 'toolkits', 'toolbox', 'toolboxes'],
+      },
       { name: 'folder-helper', folderNames: ['helpers', 'helper'] },
       { name: 'folder-serverless', folderNames: ['.serverless', 'serverless'] },
       { name: 'folder-api', folderNames: ['api', 'apis', 'restapi'] },
@@ -624,6 +635,8 @@ export const folderIcons: FolderTheme[] = [
       {
         name: 'folder-mock',
         folderNames: [
+          '_draft',
+          '_drafts',
           'mock',
           'mocks',
           'fixture',
@@ -704,7 +717,7 @@ export const folderIcons: FolderTheme[] = [
         folderNames: ['.azure-pipelines', '.azure-pipelines-ci'],
       },
       { name: 'folder-mjml', folderNames: ['mjml'] },
-      { name: 'folder-admin', folderNames: ['admin'] },
+      { name: 'folder-admin', folderNames: ['admin', 'manager', 'moderator'] },
       { name: 'folder-scala', folderNames: ['scala'] },
       {
         name: 'folder-connection',
@@ -781,7 +794,7 @@ export const folderIcons: FolderTheme[] = [
       },
       {
         name: 'folder-angular',
-        folderNames: ['.angular'],
+        folderNames: ['angular', '.angular'],
       },
       {
         name: 'folder-unity',
@@ -796,9 +809,13 @@ export const folderIcons: FolderTheme[] = [
         folderNames: ['protobufs', 'proto'],
       },
       {
+        name: 'folder-gamemaker',
+        folderNames: ['gamemaker', 'gamemaker2'],
+      },
+      {
         name: 'folder-taskfile',
         folderNames: ['taskfile', 'taskfiles', '.taskfile', '.taskfiles', 'Taskfile', 'Taskfiles', '.Taskfile', '.Taskfiles'],
-      },
+      }
     ],
   },
   {