+++ CHANGELOG.md
@@ -3,6 +3,44 @@
  All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
  
+#### [v4.24.0](https://github.com/PKief/vscode-material-icon-theme/compare/v4.23.1...v4.24.0) 
+
+> February 9, 2023 
+
+- Add folder icon for "angular" folder name [`#1859`](https://github.com/PKief/vscode-material-icon-theme/pull/1859)
+- Add  icon for .git-blame-ignore-revs [`#1836`](https://github.com/PKief/vscode-material-icon-theme/pull/1836)
+- Update contributions guidelines [`#1833`](https://github.com/PKief/vscode-material-icon-theme/pull/1833)
+- Add icon for code_of_conduct [`#1759`](https://github.com/PKief/vscode-material-icon-theme/pull/1759)
+- Added icon support for typical jekyll folders [`#1832`](https://github.com/PKief/vscode-material-icon-theme/pull/1832)
+- Add .cjs and .mjs extensions to `plop` fileNames [`#1819`](https://github.com/PKief/vscode-material-icon-theme/pull/1819)
+- Add .cjs extension to `graphql.config` [`#1818`](https://github.com/PKief/vscode-material-icon-theme/pull/1818)
+- Add Toolkit (Toolkits) and Toolbox (Toolboxes) [`#1809`](https://github.com/PKief/vscode-material-icon-theme/pull/1809)
+ 
+#### [v4.23.1](https://github.com/PKief/vscode-material-icon-theme/compare/v4.23.0...v4.23.1) 
+
+> December 23, 2022 
+
+- Add support for vitest config files [`#1572`](https://github.com/PKief/vscode-material-icon-theme/pull/1572)
+- Add icons for vscode files [`#1803`](https://github.com/PKief/vscode-material-icon-theme/pull/1803)
+- Add manager / moderator folder icon [`#1802`](https://github.com/PKief/vscode-material-icon-theme/pull/1802)
+- Update release workflow [`68bc32f`](https://github.com/PKief/vscode-material-icon-theme/commit/68bc32f)
+- Update badges in Readme [`#1796`](https://github.com/PKief/vscode-material-icon-theme/pull/1796)
+- add filename .env.uat into tune icon [`#1801`](https://github.com/PKief/vscode-material-icon-theme/pull/1801)
+ 
+#### [v4.23.0](https://github.com/PKief/vscode-material-icon-theme/compare/v4.22.0...v4.23.0) 
+
+> December 15, 2022 
+
+- feat: introduce icon folder for changesets [`#1603`](https://github.com/PKief/vscode-material-icon-theme/pull/1603)
+- Add icon for craco [`#1774`](https://github.com/PKief/vscode-material-icon-theme/pull/1774)
+- Add Angular icon to ng-package.json [`d6ea9c7`](https://github.com/PKief/vscode-material-icon-theme/commit/d6ea9c7)
+- Add tsconfig.lib.prod.json icon association [`72bd904`](https://github.com/PKief/vscode-material-icon-theme/commit/72bd904)
+- feat: support icns file type [`#1777`](https://github.com/PKief/vscode-material-icon-theme/pull/1777)
+- Add tsconfig.worker.json icon association [`90bfc9e`](https://github.com/PKief/vscode-material-icon-theme/commit/90bfc9e)
+- feat: support `etc` folder [`#1762`](https://github.com/PKief/vscode-material-icon-theme/pull/1762)
+- Add tsconfig file names [`#1756`](https://github.com/PKief/vscode-material-icon-theme/pull/1756)
+- feat: 🎨 add console icon for nushell [`#1761`](https://github.com/PKief/vscode-material-icon-theme/pull/1761)
+ 
 #### [v4.22.0](https://github.com/PKief/vscode-material-icon-theme/compare/v4.21.0...v4.22.0) 
 
 > November 5, 2022